+++ src/CurrencyLoader.php
@@ -6,35 +6,35 @@ namespace Rinvex\Country;
 
 class CurrencyLoader
 {
-    protected static $curriencies = [];
+    protected static $currencies = [];
 
     /**
-     * Retrive all the curriencies of all countries.
+     * Retrive all the currencies of all countries.
      *
-     * @param bool $longlist states if need all the details of the curriencies or only the keys
+     * @param bool $longlist states if need all the details of the currencies or only the keys
      *
      * @throws \Rinvex\Country\CountryLoaderException
      *
      * @return array
      */
-    public static function curriencies($longlist = false): array
+    public static function currencies($longlist = false): array
     {
         $list = $longlist ? 'longlist' : 'shortlist';
 
-        if (! isset(static::$curriencies[$list])) {
+        if (! isset(static::$currencies[$list])) {
             $countries = CountryLoader::countries($longlist);
 
             foreach ($countries as $country) {
                 if ($longlist) {
                     foreach ($country['currency'] as $currency => $details) {
-                        static::$curriencies[$list][$currency] = $longlist ? $details : $currency;
+                        static::$currencies[$list][$currency] = $longlist ? $details : $currency;
                     }
                 } else {
-                    static::$curriencies[$list][] = $country['currency'];
+                    static::$currencies[$list][] = $country['currency'];
                 }
             }
         }
 
-        return static::$curriencies[$list];
+        return static::$currencies[$list];
     }
 }