+++ src/Providers/CountryServiceProvider.php
@@ -17,6 +17,11 @@ class CountryServiceProvider extends ServiceProvider
         // Add country validation rule
         Validator::extend('country', function ($attribute, $value) {
             return array_key_exists(mb_strtolower($value), countries());
-        }, 'Country MUST be valid!');
+        }, trans('validation.invalid_country'));
+
+        // Add currency validation rule
+        Validator::extend('currency', function ($attribute, $value) {
+            return array_key_exists(mb_strtolower($value), currencies());
+        }, trans('validation.invalid_currency'));
     }
 }