+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v7.1.0] - 2020-07-16
+- Add currency Laravel validation rule
+- Fix wrong currencies method naming
+
 ## [v7.0.1] - 2020-04-04
 - Drop laravel/helpers usage as it's no longer used
 
@@ -117,6 +121,7 @@ This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 ## v1.0.0 - 2016-08-20
 - Tag first release.
 
+[v7.1.0]: https://github.com/rinvex/countries/compare/v7.0.1...v7.1.0
 [v7.0.1]: https://github.com/rinvex/countries/compare/v7.0.0...v7.0.1
 [v7.0.0]: https://github.com/rinvex/countries/compare/v6.1.2...v7.0.0
 [v6.1.2]: https://github.com/rinvex/countries/compare/v6.1.1...v6.1.2